+++ server/serverjs/cubefn.js
@@ -18,8 +18,10 @@
  */
 import winston from '@cubeartisan/server/serverjs/winstonConfig.js';
 
+import axios from 'axios';
 import NodeCache from 'node-cache';
 import Papa from 'papaparse';
+import sharp from 'sharp';
 
 import CardRating from '@cubeartisan/server/models/cardrating.js';
 import Cube from '@cubeartisan/server/models/cube.js';
@@ -544,6 +546,27 @@ export const saveDraftAnalytics = async (draft, seatNumber, carddb) => {
   } catch (err) {
     winston.error(err);
   }
+};
+
+const loadImage = async (url) => {
+  return (await axios({ url, responseType: 'arraybuffer' })).data;
+};
+
+export const generateSamplepackImageSharp = async (sources = [], options = {}) => {
+  const imageSources = await Promise.all(
+    sources.map(async (srcObj) => ({ ...srcObj, src: await loadImage(srcObj.src) })),
+  );
+  return sharp({
+    create: {
+      width: options.width,
+      height: options.height,
+      channels: 3,
+      background: { r: 0, g: 0, b: 0 },
+    },
+  })
+    .composite(imageSources.map(({ src, x, y }) => ({ input: src, top: y, left: x })))
+    .png()
+    .toBuffer();
 };
 
 /*